+++ core/src/main/java/org/seedstack/seed/core/internal/configuration/ClassConfigurationMapper.java
@@ -10,7 +10,6 @@ package org.seedstack.seed.core.internal.configuration;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.seedstack.coffig.TreeNode;
 import org.seedstack.coffig.spi.ConfigurationMapper;
-import org.seedstack.coffig.util.Utils;
 import org.seedstack.seed.ClassConfiguration;
 import org.seedstack.seed.SeedException;
 import org.seedstack.seed.core.internal.CoreErrorCode;
@@ -19,6 +18,8 @@ import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.HashMap;
 import java.util.Map;
+
+import static org.seedstack.shed.reflect.Types.rawClassOf;
 
 public class ClassConfigurationMapper implements ConfigurationMapper {
     @Override
@@ -35,7 +36,7 @@ public class ClassConfigurationMapper implements ConfigurationMapper {
     @Override
     @SuppressFBWarnings(value = "BC_UNCONFIRMED_CAST", justification = "Cast is verified in canHandle() method")
     public Object map(TreeNode treeNode, Type type) {
-        Class<?> rawType = Utils.getRawClass(((ParameterizedType) type).getActualTypeArguments()[0]);
+        Class<?> rawType = rawClassOf(((ParameterizedType) type).getActualTypeArguments()[0]);
         if (treeNode.type() == TreeNode.Type.MAP_NODE) {
             Map<String, String> result = new HashMap<>();
             treeNode.namedNodes()