+++ core/src/main/java/org/seedstack/seed/core/internal/lifecycle/AutoCloseableMatcher.java
@@ -12,6 +12,7 @@ import com.google.inject.Scope;
 import com.google.inject.Scopes;
 import com.google.inject.matcher.AbstractMatcher;
 import com.google.inject.spi.BindingScopingVisitor;
+import org.seedstack.seed.Ignore;
 
 import javax.inject.Singleton;
 import java.lang.annotation.Annotation;
@@ -21,7 +22,8 @@ class AutoCloseableMatcher extends AbstractMatcher<Binding<?>> {
 
     @Override
     public boolean matches(Binding<?> binding) {
-        return AutoCloseable.class.isAssignableFrom(binding.getKey().getTypeLiteral().getRawType()) && binding.acceptScopingVisitor(autoCloseableScopingVisitor);
+        Class<?> rawType = binding.getKey().getTypeLiteral().getRawType();
+        return !rawType.isAnnotationPresent(Ignore.class) && AutoCloseable.class.isAssignableFrom(rawType) && binding.acceptScopingVisitor(autoCloseableScopingVisitor);
     }
 
     private static class AutoCloseableScopingVisitor implements BindingScopingVisitor<Boolean> {